requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
coloredlogs==15.0.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
fire==0.4.0
fonttools==4.38.0
frozenlist==1.3.1
fsspec==2022.10.0
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
-e git+ssh://git@github.com/elifcansuyildiz/MasterThesis.git@cdc369db956914aed4b3af12cd281fe186454946#egg=Medical_AI
multidict==6.0.2
nest-asyncio==1.5.6
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
pluggy==1.0.0
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytorch-lightning==1.8.0.post1
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
sklearn==0.0
stack-data==0.6.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.0
torch-tb-profiler==0.4.0
torchaudio==0.13.0
torchinfo==1.7.1
torchmetrics==0.10.2
torchsampler==0.1.2
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.10.0
optuna
opencv-python
torchcam
wandb
typer
